adds sad path to photo post test

+++ app/controllers/api/v1/photos_controller.rb
@@ -13,27 +13,31 @@ class Api::V1::PhotosController < ApiController
 
 	def create
 		@event = Event.find(params[:event_id])
-
-		tempfile = Tempfile.new("fileupload")
-		begin
-			tempfile.binmode
-			tempfile.write(Base64.decode64(params[:image]))
-			filename = "#{params[:name].gsub(/\s+/, "").downcase}.jpg"
-			uploaded_file = ActionDispatch::Http::UploadedFile.new(:tempfile => tempfile, :filename => filename, :original_filename => "file")
+		if params[:image]
+			tempfile = Tempfile.new("fileupload")
+			begin
+				tempfile.binmode
+				tempfile.write(Base64.decode64(params[:image]))
+				filename = "#{params[:name].gsub(/\s+/, "").downcase}.jpg"
+				uploaded_file = ActionDispatch::Http::UploadedFile.new(:tempfile => tempfile, :filename => filename, :original_filename => "file")
+				
+				@photo = @event.photos.build(name: params[:name], image: uploaded_file)
 			
-			@photo = @event.photos.build(name: params[:name], image: uploaded_file)
-
-			if @photo.save
-				render
-			else
-				render json: {
-	    			message: 'Validation Failed',
-	    			errors: @event.errors.full_messages
-	  			}, status: 422	
+			ensure
+				tempfile.close
+				tempfile.unlink
 			end
-		ensure
-			tempfile.close
-			tempfile.unlink
+		else
+			@photo = @event.photos.build(params[:name])
+		end
+		
+		if @photo.save
+			render
+		else
+			render json: {
+		   		message: 'Validation Failed',
+		   		errors: @photo.errors.full_messages
+			}, status: 422	
 		end
 	end
